refactor: decompose AuthRoutes into domain-focused route classes
Extract PasswordRoutes, ProfileRoutes, and ApiKeyRoutes from AuthRoutes.
AuthRoutes now handles only the public auth flow (login/register/recover).
No behavior change.
